+++ src/app/helpers/debugShouldApply.js
@@ -6,7 +6,7 @@ const debugShouldApply = (tasks, configValue) => {
     return true;
   }
   const taskList = !Array.isArray(tasks) ? [tasks] : tasks;
-  const ids = [].concat(taskList.map((t) => t.id), taskList.map((t) => t.name));
+  const ids = [].concat(taskList.map((t) => (typeof t.id === 'number' ? t.id : t.type)), taskList.map((t) => (t.name ? t.name : t.type)));
 
   if (Array.isArray(configValue)) {
     return configValue.some((v) => ids.includes(v));